meta: Added support for media
The KAWCCO website now supports images and theoretical support for audio and video!
This is implemented identically to CSS in my site with 11ty’s "Passthrough File Copy" feature.
Images and other media go in the media folder, nested in folders that match the path of the
page they’re featured in.